+++ heapC.h
@@ -16,17 +16,17 @@
 //#define heap_empty(h) (!(h)->num_entries)
 
 struct heap {
-        uint32_t max_entries;
-        uint32_t num_entries;
+        uint64_t max_entries;
+        uint64_t num_entries;
         int64_t *entries; 
 };
 
-extern int32_t heap_empty(struct heap *heap);
-extern struct heap *heap_create(uint32_t max_entries);
+extern int64_t heap_empty(struct heap *heap);
+extern struct heap *heap_create(uint64_t max_entries);
 extern void heap_destroy(struct heap *heap);
 extern void heapify(struct heap *heap, int64_t i);
-extern int32_t heap_expand(struct heap *heap);
-extern int32_t heap_insert(struct heap *heap, int64_t entry);
+extern int64_t heap_expand(struct heap *heap);
+extern int64_t heap_insert(struct heap *heap, int64_t entry);
 extern int64_t heap_extract(struct heap *heap);
 extern int64_t heap_peek(struct heap *heap); 
 extern void heap_clean(struct heap *heap);
